# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |
    API specification for WASAText messaging application.
    Built according to the PDF specification.
  version: "1.0.0"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Use the user identifier returned from doLogin
  schemas:
    User:
      type: object
      description: Representation of a user in the system
      properties:
        identifier:
          type: string
          description: Unique user identifier (UUID hex string)
          example: "abcdef012345"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        name:
          type: string
          description: Username chosen by the user (3-16 characters)
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_-]+$'
          example: "Maria"
        photo:
          type: string
          format: binary
          description: User profile photo in binary format (optional)
...
    Message:
      type: object
      description: A single message within a conversation
      properties:
        messageId:
          type: string
          description: Unique message identifier
          example: "msg123"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        senderId:
          type: string
          description: User identifier of the sender
          example: "abcdef012345"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        senderName:
          type: string
          description: Username of the sender
          example: "Maria"
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_-]+$'
        content:
          type: string
          description: Text content of the message
          example: "Hello!"
          minLength: 0
          maxLength: 10000
          pattern: '^[\s\S]*$'
        photo:
          type: string
          format: binary
          description: Photo or GIF data if the message contains media
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 date and time when the message was sent
          example: "2024-01-15T10:30:00Z"
          minLength: 20
          maxLength: 30
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}'
        status:
          type: string
          enum: [sent, received, read]
          description: |
            Current delivery and read status:
            - sent: message successfully reached the server
            - received: message delivered to the recipient
            - read: recipient opened the conversation
          minLength: 4
          maxLength: 8
          pattern: '^(sent|received|read)$'
        replyTo:
          type: string
          description: Message ID of the parent message if this is a reply (optional)
          example: "msg100"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        comments:
          type: array
          minItems: 0
          maxItems: 1000
          items:
            $ref: '#/components/schemas/Comment'
          description: List of reactions/emoticons added to this message
    Comment:
      type: object
      description: A reaction to a message using an emoticon
      properties:
        userId:
          type: string
          description: User identifier of the person who reacted
          example: "abcdef012345"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        userName:
          type: string
          description: Username of the person who reacted
          example: "Maria"
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_-]+$'
        emoticon:
          type: string
          description: The single emoticon character used as a reaction
          example: "👍"
          minLength: 1
          maxLength: 8
          pattern: '^[\s\S]+$'
...
    Conversation:
      type: object
      description: Complete details of a conversation thread including all history
      properties:
        conversationId:
          type: string
          description: Unique conversation identifier
          example: "conv123"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        isGroup:
          type: boolean
          description: True for group conversations
        name:
          type: string
          description: Name of the conversation
          example: "Maria"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9 _-]+$'
        photo:
          type: string
          format: binary
          description: Image representing the conversation
        members:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/User'
          description: List of users in this conversation (for groups)
        messages:
          type: array
          minItems: 0
          maxItems: 5000
          items:
            $ref: '#/components/schemas/Message'
          description: Ordered list of messages in the conversation (newest first)
    Error:
      type: object
      description: Standard error response object
      properties:
        message:
          type: string
          description: Human-readable explanation of what went wrong
          example: "Username already taken"
          minLength: 1
          maxLength: 256
          pattern: '^[\s\S]+$'
  parameters:
...
    ConversationId:
      name: conversationId
      in: path
      description: Unique conversation identifier
      required: true
      schema:
        type: string
        description: Conversation identifier value
        minLength: 1
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
    MessageId:
      name: messageId
      in: path
      description: Unique message identifier
      required: true
      schema:
        type: string
        description: Message identifier value
        minLength: 1
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
...
paths:
...
  /conversations/{conversationId}:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    get:
      tags: ["conversation"]
      summary: Get a specific conversation with all messages
      description: |
        Opens a conversation to view all exchanged messages,
        displayed in reverse chronological order.
      operationId: getConversation
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    post:
      tags: ["message"]
      summary: Send a new message
      description: |
        Send a new message in a conversation. Can be text or photo/GIF.
        Can optionally be a reply to an existing message.
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        description: The content of the message
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Text message request
              properties:
                content:
                  type: string
                  description: Text content of the message
                  example: "Hello!"
                  minLength: 1
                  maxLength: 10000
                  pattern: '^[\s\S]+$'
                replyTo:
                  type: string
                  description: Message ID to reply to (optional)
                  example: "msg100"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
          multipart/form-data:
            schema:
              type: object
              description: Media message request
              properties:
                photo:
                  type: string
                  format: binary
                  description: Photo or GIF data
                replyTo:
                  type: string
                  description: Message ID to reply to (optional)
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages/{messageId}/forward:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/MessageId'
    post:
      tags: ["message"]
      summary: Forward a message to another conversation
      description: Forward an existing message to another conversation.
      operationId: forwardMessage
      security:
        - bearerAuth: []
      requestBody:
        description: Destination of the forwarded message
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Forward request
              properties:
                targetConversationId:
                  type: string
                  description: Conversation ID where the message should be sent
                  example: "conv456"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
              required:
                - targetConversationId
      responses:
        '201':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation or message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/MessageId'
    delete:
      tags: ["message"]
      summary: Delete a sent message
      description: |
        Delete a message that was sent by the current user.
      operationId: deleteMessage
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Message deleted successfully
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (not your message)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation or message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages/{messageId}/comments:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/MessageId'
    post:
      tags: ["comment"]
      summary: Add a reaction (comment) to a message
      description: React to a message with an emoticon.
      operationId: commentMessage
      security:
        - bearerAuth: []
      requestBody:
        description: The emoticon for the reaction
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Reaction request body
              properties:
                emoticon:
                  type: string
                  description: Single emoji/emoticon character
                  example: "👍"
                  minLength: 1
                  maxLength: 8
                  pattern: '^[\s\S]+$'
              required:
                - emoticon
      responses:
        '201':
          description: Reaction added successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation or message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: ["comment"]
      summary: Remove a reaction (uncomment) from a message
      description: Remove your own reaction from a message.
      operationId: uncommentMessage
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Reaction removed successfully
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...